TRI-32528 Remove Route Options on Prod
