+++ web/frontend/src/pages/Home/Home.module.css
@@ -292,6 +292,7 @@
 }
 .advancedoptionscontent{
     display: flex;
+    padding-top: 12px;
     flex-direction: column;
     align-items: flex-start;
     gap:24px;
@@ -314,7 +315,6 @@
     flex-direction: row;
     align-items: flex-start;
     gap:15px;
-    margin-top: 24px;
 }
 .custumradio {
     height: 18px;
